Refactor environment setup utilities and update imports
- Moved `setupEnv` function from `container.ts` to `cli.ts` for better organization of CLI-related utilities.
- Updated references to `asEnv` to `asRepoEnvironment` in multiple files to reflect the new normalization function for repository environments.
- Cleaned up imports and removed deprecated functions to streamline the codebase.
