+++ lib/types/index.ts
@@ -234,6 +234,17 @@ export type RepoEnvironment =
   | { kind: "host"; root: string }
   | { kind: "container"; name: string; mount?: string }
 
+/**
+ * Helper to normalize legacy baseDir string to RepoEnvironment
+ */
+export function asRepoEnvironment(
+  arg: string | RepoEnvironment
+): RepoEnvironment {
+  return typeof arg === "string"
+    ? { kind: "host", root: arg } // auto-wrap legacy baseDir
+    : arg
+}
+
 // Type exports
 export type AnyEvent = z.infer<typeof anyEventSchema>
 export type BaseEvent = z.infer<typeof baseEventSchema>
